requirements.txt
anyio==3.6.2
APScheduler==3.10.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
ephem==4.1.4
h11==0.14.0
h2==4.1.0
h3==3.7.6
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
hyperframe==6.0.1
idna==3.4
numpy==1.24.2
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.5
pyephem==9.99
python-dateutil==2.8.2
python-dotenv==0.21.1
python-i18n==0.3.9
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
timezonefinder==6.1.9
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
